+++ src/Feliz/Helper/Helpers.fs
@@ -11,23 +11,19 @@ module Helpers =
     let optDispose (disposeOption: #IDisposable option) =
         { new IDisposable with member _.Dispose () = disposeOption |> Option.iter (fun d -> d.Dispose()) }
 
-
 [<RequireQualifiedAccess>]
 module Interop =
 
     open Fable.Core
     open Fable.React
 
-    [<Emit "undefined">]
-    let undefined: obj = jsNative
+    [<Emit("($0 !== null && typeof $0 === \"object\" && !Array.isArray($0))")>]
+    let isObject (value: obj) = jsNative
 
-    let inline svgAttribute (key: string) (value: obj) : ISvgAttribute = unbox (key, value)
+    let setKeyOnObj (withKey: ('props -> string) option) props =
+        match withKey with
+        | Some f ->
+            props?key <- f props
+            props
+        | None -> props
 
-    [<Emit "typeof $0 === 'number'">]
-    let isTypeofNumber (x: obj) : bool = jsNative
-
-    [<Emit("typeof $0[Symbol.iterator] === 'function'")>]
-    let isIterable (x: obj): bool = jsNative
-
-    [<Emit("""typeof $0 === "string" || $0 instanceof String""")>]
-    let isString (value: obj): bool = jsNative