+++ src/Feliz.CompilerPlugins/CHANGELOG.md
@@ -16,6 +16,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## 3.0.0-rc.7 - 2025-11-26
+
+### 🗑️ Deprecated
+
+- Removed name setting for memo components, as this would remove the `memo` tag in react dev tools (by @Freymaurer)
+
 ## 3.0.0-rc.6 - 2025-11-21
 
 ### 🐛 Fixed